+++ api/rust/cargo/lief/src/elf/binary.rs
@@ -1,4 +1,5 @@
 use std::mem::size_of;
+use std::pin::Pin;
 
 use num_traits::{Num, cast};
 
@@ -291,6 +292,16 @@ impl Binary {
 impl generic::Binary for Binary {
     fn as_generic(&self) -> &ffi::AbstractBinary {
         self.ptr.as_ref().unwrap().as_ref()
+    }
+
+    fn as_pin_mut_generic(&mut self) -> Pin<&mut ffi::AbstractBinary> {
+        unsafe {
+            Pin::new_unchecked({
+                (self.ptr.as_ref().unwrap().as_ref()
+                    as *const ffi::AbstractBinary
+                    as *mut ffi::AbstractBinary).as_mut().unwrap()
+            })
+        }
     }
 }